added job parameter access
git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@474 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ src/main/java/de/cebitec/mgx/client/upload/CallbackI.java
@@ -6,5 +6,5 @@ package de.cebitec.mgx.client.upload;
  */
 public interface CallbackI {
 
-    public void callback(int i);
+    public void callback(long i);
 }
